ALL the commands
Literally! The port is complete! But untested. woooo

+++ src/org/usfirst/frc2175/Robot2014/subsystems/Arm.java
@@ -49,7 +49,7 @@ public class Arm extends Subsystem {
 	}
     }
     
-    public void RunRollerBarAtSpeed(float speed) {
+    public void RunRollerBarAtSpeed(double speed) {
         rollerBar.set(speed);
     }
